add continuous autotrans for replay + 2 bug fix
- bug 1: Common.twait was set at 0.02 instead of 0.08
- bug 2: integrate_and_saturate was wrong: it was integrating its input then computing the clamped result from it. The problem was that when the derivative dx became positive, the output did not start to climb immediately because the integrated signal was probably way below the minimal threshold.
Now in autotransc the node is an automaton with 3 modes: High if x >= xmax until dx < 0, Low if x <= xmin until dx > 0, and Integrate if x in ]xmin,xmax[ until x >= xmin or x <= xmax

Possible issue: when a transition is triggered from Integrate, it seems like the value of x is equal to either xmin or xmax, which is correct. But it is not enforced in the model and that value is whatever the value of x is at te moment where the zero-crossing is triggered. x is obtained by integrating a signal so it is continuous, as long as xmin and xmax are too.
This might need to be enforced though.