bug fix
the interpolation function for AFC was outputting a piecewise constant signal for input engine instead of a constant one.
the computations in zelus were correct though because the nodes in afc_bench were only reading the first value of the input. However, the dump was written with the piecewise constant signal wich made it more difficult to feed it to matlab to validate the results.